switch to float for sequence identity of mapped reads for better resolution of differences

+++ MGX-API/src/main/java/de/cebitec/mgx/api/model/MappedSequenceI.java
@@ -23,7 +23,7 @@ public abstract class MappedSequenceI extends LocationBase<MappedSequenceI> {
 
     public abstract long getSeqId();
 
-    public abstract int getIdentity();
+    public abstract float getIdentity();
 
     @Override
     public abstract int compareTo(MappedSequenceI o);
